Introduce new class for WS proxy configs
